table::column_operator

Trait ColumnBooleanOperator

pub trait ColumnBooleanOperator<T, U> {

    // Required methods
    fn equal<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn eq_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn eq_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn eq_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn neq<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn neq_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn neq_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn neq_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn gt<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn gt_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn gt_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn gt_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn ge<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn ge_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn ge_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn ge_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn lt<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn lt_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn lt_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn lt_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn le<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn le_binary<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn le_public<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn le_public_binary<N: Network>(
        &self,
        pub_element: &U,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn in_public_binary<N: Network>(
        &self,
        pub_elements: &Vec<U>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn and<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;
    fn or<N: Network>(
        &self,
        other_column: &ShareColumn<T>,
        netstate_args: &mut NetStateArgs<'_, N>,
    ) -> Result<ShareColumn<T>>;

}

Required Methods

fn equal<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

_binary means the input is binary shares and does not require transform from arithmetic to binary

fn eq_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn eq_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

Compare two columns and return a binary T share result

fn eq_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn neq<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn neq_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn neq_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn neq_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn gt<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn gt_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn gt_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn gt_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn ge<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn ge_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn ge_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn ge_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn lt<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn lt_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn lt_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn lt_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn le<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn le_binary<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn le_public<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn le_public_binary<N: Network>( &self, pub_element: &U, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn in_public_binary<N: Network>( &self, pub_elements: &Vec<U>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

Compare two columns and return a binary T share result

fn and<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

fn or<N: Network>( &self, other_column: &ShareColumn<T>, netstate_args: &mut NetStateArgs<'_, N>, ) -> Result<ShareColumn<T>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.