table
Module predicate

Macros

return_none_if_utf8 🔒

Assert the scalar value is not utf8. Returns None if it’s utf8. In theory, it should be converted to a timestamp scalar value by TypeConversionRule.

Structs

Predicate

Reference-counted pointer to a list of logical exprs.

Functions

build_time_range_predicate

build_time_range_predicate extracts time range from logical exprs to facilitate fast time range pruning.

extract_from_between_expr 🔒

extract_from_binary_expr 🔒

extract_from_in_list_expr 🔒

Extract time range filter from IN (...) expr.

extract_time_range_from_expr 🔒

Extract time range filter from WHERE/IN (...)/BETWEEN clauses. Return None if no time range can be found in expr.

get_timestamp_filter 🔒