Trait table::table::Table

pub trait Table: Send + Sync {

    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn schema(&self) -> SchemaRef;
    fn table_info(&self) -> TableInfoRef;
    fn scan_to_stream<'life0, 'async_trait>(
        &'life0 self,
        request: ScanRequest
    ) -> Pin<Box<dyn Future<Output = Result<SendableRecordBatchStream>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn table_type(&self) -> TableType { ... }
    fn insert<'life0, 'async_trait>(
        &'life0 self,
        _request: InsertRequest
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn supports_filters_pushdown(
        &self,
        filters: &[&Expr]
    ) -> Result<Vec<FilterPushDownType>> { ... }
    fn alter<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _context: AlterContext,
        _request: &'life1 AlterTableRequest
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn delete<'life0, 'async_trait>(
        &'life0 self,
        _request: DeleteRequest
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn flush<'life0, 'async_trait>(
        &'life0 self,
        region_number: Option<RegionNumber>,
        wait: Option<bool>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn close<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _regions: &'life1 [RegionNumber]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn region_stats(&self) -> Result<Vec<RegionStat>> { ... }
    fn contains_region(&self, _region: RegionNumber) -> Result<bool> { ... }
    fn statistics(&self) -> Option<TableStatistics> { ... }
    fn compact<'life0, 'async_trait>(
        &'life0 self,
        region_number: Option<RegionNumber>,
        wait: Option<bool>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn truncate<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }

}

Table abstraction.

Required Methods

fn as_any(&self) -> &dyn Any

Returns the table as Any so that it can be downcast to a specific implementation.

fn schema(&self) -> SchemaRef

Get a reference to the schema for this table

fn table_info(&self) -> TableInfoRef

Get a reference to the table info.

fn scan_to_stream<'life0, 'async_trait>( &'life0 self, request: ScanRequest ) -> Pin<Box<dyn Future<Output = Result<SendableRecordBatchStream>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn table_type(&self) -> TableType

Get the type of this table for metadata/catalog purposes.

fn insert<'life0, 'async_trait>( &'life0 self, _request: InsertRequest ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Insert values into table.

Returns number of inserted rows.

fn supports_filters_pushdown( &self, filters: &[&Expr] ) -> Result<Vec<FilterPushDownType>>

Tests whether the table provider can make use of any or all filter expressions to optimise data retrieval.

fn alter<'life0, 'life1, 'async_trait>( &'life0 self, _context: AlterContext, _request: &'life1 AlterTableRequest ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Alter table.

fn delete<'life0, 'async_trait>( &'life0 self, _request: DeleteRequest ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Delete rows in the table.

Returns number of deleted rows.

fn flush<'life0, 'async_trait>( &'life0 self, region_number: Option<RegionNumber>, wait: Option<bool> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Flush table.

Options:

region_number: specify region to flush.

wait: Whether to wait until flush is done.

fn close<'life0, 'life1, 'async_trait>( &'life0 self, _regions: &'life1 [RegionNumber] ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Close the table.

fn region_stats(&self) -> Result<Vec<RegionStat>>

Get region stats in this table.

fn contains_region(&self, _region: RegionNumber) -> Result<bool>

Return true if contains the region

fn statistics(&self) -> Option<TableStatistics>

Get statistics for this table, if available

fn compact<'life0, 'async_trait>( &'life0 self, region_number: Option<RegionNumber>, wait: Option<bool> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn truncate<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,