Module table::metadata

Structs

RawTableInfo

Struct used to serialize and deserialize TableInfo.

RawTableMeta

Struct used to serialize and deserialize TableMeta.

SplitResult 🔒

The result after splitting requests by column location info.

TableIdent

Identifier of the table.

TableInfo

TableInfoBuilder

Builder for TableInfo.

TableMeta

The table metadata Note: if you add new fields to this struct, please ensure ‘new_meta_builder’ function works. TODO(dennis): find a better way to ensure ‘new_meta_builder’ works when adding new fields.

TableMetaBuilder

Builder for TableMeta.

Enums

ConvertError

FilterPushDownType

Indicates whether and how a filter expression can be handled by a Table for table scans.

TableInfoBuilderError

Error type for TableInfoBuilder

TableMetaBuilderError

Error type for TableMetaBuilder

TableType

Indicates the type of this table for metadata/catalog purposes.

Type Aliases

ConvertResult

TableId

TableInfoRef

TableVersion