table

Module metadata

Structs

RawTableInfo

Struct used to serialize and deserialize TableInfo.

RawTableMeta

Struct used to serialize and deserialize TableMeta.

SplitResult 🔒

The result after splitting requests by column location info.

TableIdent

Identifier of the table.

TableInfo

TableInfoBuilder

Builder for TableInfo.

TableMeta

The table metadata.

TableMetaBuilder

Builder for TableMeta.

Enums

ConvertError

FilterPushDownType

Indicates whether and how a filter expression can be handled by a Table for table scans.

TableInfoBuilderError

Error type for TableInfoBuilder

TableMetaBuilderError

Error type for TableMetaBuilder

TableType

Indicates the type of this table for metadata/catalog purposes.

Functions

set_column_fulltext_options 🔒

Set column fulltext options if it passed the validation.

set_column_skipping_index_options 🔒

unset_column_fulltext_options 🔒

unset_column_skipping_index_options 🔒

Type Aliases

ConvertResult

TableId

TableInfoRef

TableVersion