table

Module requests

Table and TableEngine requests

Structs

AddColumnRequest

Add column request

AlterTableRequest

Alter table request

BuildIndexTableRequest

CompactTableRequest

CopyDatabaseRequest

CopyQueryToRequest

CopyTableRequest

Copy table request

DeleteRequest

Delete (by primary key) request

FlushTableRequest

InsertRequest

ModifyColumnTypeRequest

Change column datatype request

SetDefaultRequest

TableOptions

TruncateTableRequest

Truncate table request

Enums

AlterKind

CopyDirection

SetIndexOption

UnsetIndexOption

Constants

AUTO_CREATE_TABLE_KEY

COMMENT_KEY

FILE_TABLE_FORMAT_KEY

FILE_TABLE_LOCATION_KEY

FILE_TABLE_META_KEY

FILE_TABLE_PATTERN_KEY

OTLP_METRIC_COMPAT_KEY

OTLP_METRIC_COMPAT_PROM

SKIP_WAL_KEY

STORAGE_KEY

TABLE_DATA_MODEL

TABLE_DATA_MODEL_TRACE_V1

TTL_KEY

VALID_TABLE_OPTION_KEYS

WRITE_BUFFER_SIZE_KEY

Statics

VALID_DB_OPT_KEYS ๐

Functions

validate_database_option

Returns true if the key is a valid key for database.

validate_table_option

Returns true if the key is a valid key for any engine or storage.